Implements breathing and basic camera rotation
Breathing system allows player to press the inhalation key to breath in, release the key to begin holding the breath, press the exhalation key to breath out, and release the exhalation key to stop exhaling and gain a directional speed boost based on how accurate the player was when breathing.
Camera rotation turns the player based on the directional input, steadying to max rotation or static rotation slowly.