Revamps breath procedure timing
Whenever the player successfully or unsuccessfully finishes a breathing loop (either full or incomplete), the breathCoolDownTimer variable is set to zero. If the player starts another breathing loop before the breathCoolDownTimer is >= minStableBreathTime, they will suffer a speed decrease and breathCoolDownTimer will be set to zero again. breathlessTimer now only increases when the player is not actively attempting breathing, but it will not return to zero until a successful box-breathing loop.